python_api/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.11",
    "mysql-connector-python>=9.2.0",
    "uvicorn>=0.34.0",
]
